Rename commands to repl and add general namespace "http"
