+++ src/common/models.hpp
@@ -3,23 +3,26 @@
 #include <string>
 #include <unordered_map>
 
-namespace models
+namespace http
 {
-  struct Request
+  namespace models
   {
-    std::string method;
-    std::string host;
-    std::string path;
-    std::unordered_map< std::string, std::string > headers;
-    std::string body;
-  };
+    struct Request
+    {
+      std::string method;
+      std::string host;
+      std::string path;
+      std::unordered_map< std::string, std::string > headers;
+      std::string body;
+    };
 
-  struct Response
-  {
-    std::string status_line;
-    std::unordered_map< std::string, std::string > headers;
-    std::string body;
-    std::string error;
-  };
-}
+    struct Response
+    {
+      std::string status_line;
+      std::unordered_map< std::string, std::string > headers;
+      std::string body;
+      std::string error;
+    };
+  }
 #endif
+}
