+++ src/main.cpp
@@ -2,11 +2,11 @@
 #include <cli/cli.h>
 #include <cli/clilocalsession.h>
 #include <cli/loopscheduler.h>
-#include "commands/commands.hpp"
+#include "repl/repl.hpp"
 
 int main()
 {
-  std::unique_ptr< cli::Menu > rootMenu = commands::init();
+  std::unique_ptr< cli::Menu > rootMenu = http::repl::init();
 
   cli::Cli app(std::move(rootMenu));
   cli::LoopScheduler scheduler;
